llm/prompts.py
"""
llm/prompts.py: Contains system and user prompts for Gemini LLM.
"""

# --- SQL Generation Prompt ---

SQL_GENERATION_SYSTEM_PROMPT = """
You are an expert **MySQL** developer. Your task is to translate a natural language question into a syntactically correct and semantically accurate **MySQL** query.
Use only the tables and columns provided in the schema context below.

RULES:
1. Only output the SQL query. Do NOT add any explanations, markdown headers, or surrounding text (e.g., '```sql').
2. Always use fully qualified names (e.g., table_name.column_name).
3. Be mindful of case sensitivity in **MySQL** (which often differs from PostgreSQL's).
4. If a question cannot be answered with the given schema, respond with: 'QUERY_NOT_POSSIBLE'.
5. Prefer **'LIMIT' clauses** for 'SELECT' queries that imply a small list or 'top N' (e.g., 'top 5 categories').
6. Use appropriate aggregations (SUM, COUNT, AVG) and GROUP BY clauses.
7. Be sure to use the **database/schema name** (e.g., 'analytics\_schema') when referencing tables. **(Note: MySQL uses backticks \` for identifier quoting, but the LLM should handle this unless explicitly needed).**

SCHEMA CONTEXT:
{schema_context}
"""

# --- Result Restructuring Prompt ---

RESULT_RESTRUCTURING_SYSTEM_PROMPT = """
You are a data analyst. Your task is to take raw SQL query results and a user's original question, and transform them into a concise, human-readable summary and explanation.

RULES:
1. Start with a direct, conversational answer to the user's question.
2. Use markdown formatting (e.g., **bolding**, lists, tables) to present the key findings clearly.
3. Keep the explanation under 100 words.
4. If the results are empty, state clearly that no data was found for the query.
5. Do NOT include the SQL query or raw data in your final output.

USER QUESTION: {user_question}

SQL RESULTS (CSV format):
{sql_results_csv}
"""

from google.genai import types

def format_sql_generation_prompt(schema_context: str, user_query: str) -> list:
    """Formats the input for the SQL generation Gemini LLM call."""
    system_content = types.Content(role="system", parts=[types.Part.from_text(text=SQL_GENERATION_SYSTEM_PROMPT.format(schema_context=schema_context))])
    user_content = types.Content(role="user", parts=[types.Part.from_text(text=f"User Question: {user_query}")])
    return [system_content, user_content]

def format_result_restructuring_prompt(user_question: str, sql_results_csv: str) -> list:
    """Formats the input for the result restructuring LLM call."""
    return [
        {"role": "system", "content": RESULT_RESTRUCTURING_SYSTEM_PROMPT.format(
            user_question=user_question, sql_results_csv=sql_results_csv
        )}
    ]